# src/main/java/com/priyhotel/entity/Hotel.java
package com.priyhotel.entity;

import jakarta.persistence.*;
import lombok.Data;

import java.util.ArrayList;
import java.util.List;

@Data
@Entity
@Table(name = "hotels")
public class Hotel {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false)
    private String name;

    @ManyToOne
    @JoinColumn(name = "location_id", nullable = false)
    private Location location;

    @Column(nullable = false)
    private String address;

    @Column(nullable = false)
    private String contactNumber;

    @Column(nullable = false)
    private String email;

    @OneToMany(mappedBy = "hotel", cascade = CascadeType.ALL)
    private List<Review> reviews;

    @ManyToMany
    @JoinTable(
            name = "hotel_room_types", // Join table name
            joinColumns = @JoinColumn(name = "hotel_id"), // FK to Hotel
            inverseJoinColumns = @JoinColumn(name = "room_type_id") // FK to RoomType
    )
    private List<RoomType> roomTypes;

    @ManyToMany
    @JoinTable(
            name = "hotel_assets",
            joinColumns = @JoinColumn(name = "hotel_id"),
            inverseJoinColumns = @JoinColumn(name = "asset_id")
    )
    private List<Asset> assets = new ArrayList<>();
}